Trait stylist_core::ResultDisplay

pub trait ResultDisplay<T> {
    fn unwrap_display(self) -> T;
    fn expect_display(self, msg: &str) -> T;
}

Required Methods

fn unwrap_display(self) -> T

Returns the contained Ok value, consuming the self value, panic when Result is Err.

fn expect_display(self, msg: &str) -> T

Returns the contained Ok value, consuming the self value, panic with message when Result is Err.